requirements.txt
# NoFina 项目依赖
requests>=2.31.0
pyyaml>=6.0.1
notion-client>=2.2.1
pandas>=2.0.3
python-dateutil>=2.8.2
pytz>=2023.3
schedule>=1.2.0
